tests/wpt/web-platform-tests/notifications/permission.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Notification.permission (basic)</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<link rel="author" title="Xin Liu" href="mailto:xinx.liu@intel.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
    assert_in_array(Notification.permission,
        ["default", "denied", "granted"],
        "Notification.permission")
}, "Checked the Notification.permission property.")
</script>
